DFR-1533: Tech Debt - Refactor Orchestrator service to use CaseType enum instead of constants
