+++ src/test/java/uk/gov/hmcts/reform/finrem/caseorchestration/service/ApprovedOrderNoticeOfHearingServiceTest.java
@@ -15,6 +15,7 @@ import uk.gov.hmcts.reform.finrem.caseorchestration.config.DocumentConfiguration
 import uk.gov.hmcts.reform.finrem.caseorchestration.helper.DocumentHelper;
 import uk.gov.hmcts.reform.finrem.caseorchestration.model.ccd.AdditionalHearingDocument;
 import uk.gov.hmcts.reform.finrem.caseorchestration.model.ccd.CaseDocument;
+import uk.gov.hmcts.reform.finrem.caseorchestration.model.ccd.CaseType;
 import uk.gov.hmcts.reform.finrem.caseorchestration.model.ccd.Element;
 import uk.gov.hmcts.reform.finrem.caseorchestration.model.document.BulkPrintDocument;
 import uk.gov.hmcts.reform.finrem.caseorchestration.service.noc.solicitors.CheckApplicantSolicitorIsDigitalService;
@@ -34,8 +35,6 @@ import static org.mockito.ArgumentMatchers.eq;
 import static org.mockito.Mockito.times;
 import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
-import static uk.gov.hmcts.reform.finrem.caseorchestration.OrchestrationConstants.CASE_TYPE_ID_CONSENTED;
-import static uk.gov.hmcts.reform.finrem.caseorchestration.OrchestrationConstants.CASE_TYPE_ID_CONTESTED;
 import static uk.gov.hmcts.reform.finrem.caseorchestration.OrchestrationConstants.YES_VALUE;
 import static uk.gov.hmcts.reform.finrem.caseorchestration.TestConstants.AUTH_TOKEN;
 import static uk.gov.hmcts.reform.finrem.caseorchestration.TestSetUpUtils.DOC_URL;
@@ -99,7 +98,7 @@ public class ApprovedOrderNoticeOfHearingServiceTest extends BaseServiceTest {
         when(caseDataService.isConsentedApplication(caseDetails)).thenReturn(false);
 
         caseDetails.getData().put(ANOTHER_HEARING_TO_BE_LISTED, YES_VALUE);
-        caseDetails.setCaseTypeId(CASE_TYPE_ID_CONTESTED);
+        caseDetails.setCaseTypeId(CaseType.CONTESTED.getCcdType());
         approvedOrderNoticeOfHearingService.createAndStoreHearingNoticeDocumentPack(caseDetails, AUTH_TOKEN);
 
         assertCaseDataHasHearingNoticesCollection();
@@ -117,7 +116,7 @@ public class ApprovedOrderNoticeOfHearingServiceTest extends BaseServiceTest {
         when(caseDataService.isConsentedApplication(caseDetails)).thenReturn(true);
 
         caseDetails.getData().put(ANOTHER_HEARING_TO_BE_LISTED, YES_VALUE);
-        caseDetails.setCaseTypeId(CASE_TYPE_ID_CONSENTED);
+        caseDetails.setCaseTypeId(CaseType.CONSENTED.getCcdType());
         approvedOrderNoticeOfHearingService.createAndStoreHearingNoticeDocumentPack(caseDetails, AUTH_TOKEN);
 
         assertCaseDataHasHearingNoticesCollection();
