+++ src/test/java/uk/gov/hmcts/reform/finrem/caseorchestration/service/AssignCaseAccessServiceTest.java
@@ -20,6 +20,7 @@ import uk.gov.hmcts.reform.finrem.caseorchestration.model.ccd.CaseAssignmentUser
 import uk.gov.hmcts.reform.finrem.caseorchestration.model.ccd.CaseAssignmentUserRolesRequest;
 import uk.gov.hmcts.reform.finrem.caseorchestration.model.ccd.CaseAssignmentUserRolesResource;
 import uk.gov.hmcts.reform.finrem.caseorchestration.model.ccd.CaseAssignmentUserRolesResponse;
+import uk.gov.hmcts.reform.finrem.caseorchestration.model.ccd.CaseType;
 
 import java.util.List;
 import java.util.Set;
@@ -44,7 +45,6 @@ import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
 import static org.springframework.http.HttpHeaders.CONTENT_TYPE;
 import static org.springframework.http.MediaType.APPLICATION_JSON_VALUE;
-import static uk.gov.hmcts.reform.finrem.caseorchestration.OrchestrationConstants.CASE_TYPE_ID_CONTESTED;
 import static uk.gov.hmcts.reform.finrem.caseorchestration.TestConstants.AUTH_TOKEN;
 import static uk.gov.hmcts.reform.finrem.caseorchestration.TestConstants.TEST_CASE_ID;
 import static uk.gov.hmcts.reform.finrem.caseorchestration.TestConstants.TEST_URL;
@@ -89,7 +89,7 @@ public class AssignCaseAccessServiceTest extends BaseServiceTest {
         assignCaseAccessRequest = AssignCaseAccessRequest
             .builder()
             .case_id(TEST_CASE_ID)
-            .case_type_id(CASE_TYPE_ID_CONTESTED)
+            .case_type_id(CaseType.CONTESTED.getCcdType())
             .assignee_id(TEST_USER_ID)
             .build();
